requirements.txt
einops==0.8.0
matplotlib==3.9.0
medmnist==3.0.1
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
scikit-image==0.24.0
scikit-learn==1.5.0
torchmetrics==1.4.1
tqdm==4.66.4
transformers==4.43.3
dill==0.3.8
dm-haiku==0.0.12